requirements.txt
omegaconf
discord.py
PyNaCl
